==> 01_Crawler.ipynb <==
import requests
from bs4 import BeautifulSoup
from lxml import etree
from urllib.parse import urljoin
import pandas as pd
import re
import numpy as np

# Index range for all 34 matchdays each season
matchday_ind_ls = range(1, 35)

# Index range for all season years from 1965 (1965/67) to 2022 (2022/23)
season_ind_ls = range(1963, 2023)

# Index range for all 18 clubs
clubs_ind_ls = range(1,19)


# Base URL of Transfermarkt Webpage
BASE_URL = 'https://www.transfermarkt.de/'

# header config for Browser setup
headers = {'User-Agent': 
           'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/47.0.2526.106 Safari/537.36'}


def CLUBS_OVERVIEW_URL(season):
    """
    Reads the url for overview of all clubs for specific season

    Args:
        season (int): Accept years in form of yyyy

    Returns:
        response object: Provides methods and attributes to access the data returned by the HTTP request
    """
    url = urljoin(BASE_URL, f'bundesliga/startseite/wettbewerb/L1/plus/?saison_id={season}')
    response = requests.get(url, headers=headers)
    print(url)
    return response

# Overview over all match results of each season and matchdays
def MATCH_RESULTS_URL(season, matchday):
    """
    Reads the url for all match results of a specific season and a specific matchday
    
    Args:
        season (int): Accept years in form of yyyy
        matchday (int): Accepts an index from 1 to 18 for the number of matchdays

    Returns:
        response object: Provides methods and attributes to access the data returned by the HTTP request
    """
    url = urljoin(BASE_URL, f'bundesliga/spieltag/wettbewerb/L1/plus/?saison_id={season}&spieltag={matchday}')
    response = requests.get(url, headers=headers)
    print(url)
    return response


# Overview over all match results of each season and matchdays
def TRANSFERS_URL(season):
    """
    Reads the url for all transfer amounts of a specific season
    
    Args:
        season (int): Accept years in form of yyyy

    Returns:
        response object: Provides methods and attributes to access the data returned by the HTTP request
    """
    url = urljoin(BASE_URL, f'bundesliga/transfers/wettbewerb/L1/plus/?saison_id={season}&s_w=&leihe=1&intern=0&intern=1')
    response = requests.get(url, headers=headers)
    print(url)
    return response


def dict_to_df(col_name_ls, value_ls):
    """
    Reads the column names and their values in form of lists. Store it into one dictionary to convert it to a Dataframe format 
    
    Args:
        col_name_ls (list['str']): list of column names
        value_ls (list('list')): list of lists with the values for each column

    Returns:
        df (DataFrame):  
    """
    dict = {}
    for enum in range(len(col_name_ls)):
        dict[col_name_ls[enum]] = value_ls[enum]
    df = pd.DataFrame(dict)
    return df

# Clubs Overview over all seasons

# Loop through all season for clubs overview
df_clubs_overview = pd.DataFrame()
for season in season_ind_ls:
    response = CLUBS_OVERVIEW_URL(season)
    soup = BeautifulSoup(response.content, 'lxml')

    # Clubnames 
    clubs_ls = []
    clubs = soup.find('div', id='yw1').find_all('td', class_='hauptlink no-border-links')
    for club in clubs:
        clubs_ls.append(club.text.strip())

    # The number of players in the pool
    players_count_ls = []
    players_count = soup.find('div', id='yw1').find_all('td', class_ = 'zentriert')
    for player_count in players_count:
        for item in player_count.find_all('a'):
            if item.text != '':
                players_count_ls.append(item.text.strip())

    # 1963 and 1964 has 16 clubs and season 1991 has 20 clubs. Regular 18 clubs
    clubs_ind_ls = range(1,19)
    if season == 1963 or season == 1964:
        clubs_ind_ls = range(1,17)
    if season == 1991:
        clubs_ind_ls = range(1,21)

    # Average age of all players within a club
    players_avg_age_ls = []
    for row in clubs_ind_ls:
        dom = etree.HTML(str(soup))
        value = dom.xpath(f"//*[@id='yw1']/table/tbody/tr[{row}]/td[4]")[0].text
        players_avg_age_ls.append(value.strip())
    players_avg_age_ls

    # Number of legionaries within a club
    legionaries_ls = []
    for row in clubs_ind_ls:
        dom = etree.HTML(str(soup))
        value = dom.xpath(f"//*[@id='yw1']/table/tbody/tr[{row}]/td[5]")[0].text
        legionaries_ls.append(value.strip())
    legionaries_ls

    # Average market value of each club
    avg_market_value_ls = []
    for row in clubs_ind_ls:
        dom = etree.HTML(str(soup))
        value = dom.xpath(f"//*[@id='yw1']/table/tbody/tr[{row}]/td[6]")[0].text
        avg_market_value_ls.append(value.strip())
    avg_market_value_ls

    # Total market value of each club
    total_market_values_ls =[]
    total_market_values = soup.find('table', class_='items').find_all('td', class_="rechts")
    for item1 in total_market_values:
        for item2 in item1.find_all('a'):
            total_market_values_ls.append(item2.text.strip())
    total_market_values_ls


    # Write all Information to a Dataframe
    col_name_ls = ['CLUB_NAME', 'PLAYERS_COUNT', 'PLAYERS_AVG_AGE', 'LEGIONARIES_COUNT', 'AVG_MARKET_VALUE', 'TOTAL_MARKET_VALUE']
    value_ls = [clubs_ls, players_count_ls, players_avg_age_ls, legionaries_ls, avg_market_value_ls, total_market_values_ls]
    df = dict_to_df(col_name_ls=col_name_ls, value_ls=value_ls) 

    # add season information as column to dataframe
    df['season'] = f"{season}"

    # Concat all dataframes of each season into one dataframe 
    df_clubs_overview = pd.concat([df_clubs_overview, df])

# save dataframe to pickle
#df_clubs_overview.to_pickle("data/df_clubs_overview.pkl")

df_clubs_overview.to_csv('./data/df_clubs_overview.csv', index=False)

# read data from pickle
df_clubs_overview = pd.read_pickle("./data/df_clubs_overview.pkl")

df_clubs_overview

# Retrieve final table for each season

def SEASON_TABLE_URL(season):
    """
    Reads the url for the final table for e specific season

    Args:
        season (int): Accept years in form of yyyy

    Returns:
        response object: Provides methods and attributes to access the data returned by the HTTP request
    """
    url = urljoin(BASE_URL, f'bundesliga/tabelle/wettbewerb/L1?saison_id={season}')
    response = requests.get(url, headers=headers)
    print(url)
    return response

def get_final_table(season): 
    """Retrives final table for a given season"""
    response = SEASON_TABLE_URL(season)
    soup = BeautifulSoup(response.content, 'lxml')
    
    # retrieve places and clubs
    elements = soup.find_all('td', class_="no-border-links hauptlink")
    element_list = [elem.text.strip() for elem in elements][:18]
    places = [i for i in range(1, 19)]
    
    return pd.DataFrame({"SEASON": season, "CLUB_NAME": element_list, "PLACE": places})

df_final_table = pd.DataFrame()

for season in range(2004, 2023):
    df_final_table = pd.concat([df_final_table, get_final_table(season)])

df_final_table

# save dataframe to pickle
df_final_table.to_pickle("./data/df_final_table.pkl")

# Retrieve matchday data

def extract_betting_percentages(word: str):
    """Extracts the betting percentages for win home team, remis and win away team"""
    # Define the regular expression pattern
    pattern = r'([\d,]+) %'

    # Find all matches using the pattern
    matches = re.findall(pattern, word)

    # Convert the matched values to floats
    return [float(match.replace(',', '.')) for match in matches] 

def extract_club_standings(word: str, pattern: str, home: bool):
    """Extracts club and standing from given word"""
    # Find the match using the pattern
    match = re.match(pattern, word)

    if match:
        if home:
            number = match.group(1)
            club = match.group(2)
        else:
            club = match.group(1)
            number = match.group(2)
        
        return club, number
    else:
        print("No match found.")
        return None, None

def get_matchday_data(season: int, matchday: int):
    """Retrieve matchday data for given season and matchday"""
    
    response = MATCH_RESULTS_URL(season, matchday)
    soup = BeautifulSoup(response.content, 'lxml')
    
    # retrieve dates for matchdays
    elements = soup.find_all('a')
    element_list = [elem.text.strip() for elem in elements]
    date_pattern = r'\b\d{2}\.\d{2}\.\d{4}\b'
    dates = [date for element in element_list for date in re.findall(date_pattern, element)]
    
    # retrieve tipps for matchdays
    match_tipps = soup.find_all('tr', class_='no-border tm-user-tendenz')
    tipp_list = [tipp.text.strip() for tipp in match_tipps]
    tipps = [tuple(extract_betting_percentages(tipp)) for tipp in tipp_list]

    # if no tipps exist, use nan
    if len(tipps) < 9:
        tipps=[(np.nan,np.nan,np.nan,)]*9
    if season == 1991:
        tipps=[(np.nan,np.nan,np.nan,)]*10
    
    # retrieve all home teams for given matchday
    home_teams = soup.find_all('td', class_='rechts hauptlink no-border-rechts hide-for-small spieltagsansicht-vereinsname')
    home_list = [team.text.strip() for team in home_teams]
    pattern = r'^\((\d+)\.\)\s+(.+)$'
    home = [(extract_club_standings(team, pattern, home=True)) for team in home_list]
    
    # retrieve all guest teams for given matchday
    teams_away = soup.find_all('td', class_='hauptlink no-border-links no-border-rechts hide-for-small spieltagsansicht-vereinsname')
    away_list = [team.text.strip() for team in teams_away]
  
    pattern = r'^(.+)\s+\((\d+)\.\)$'
    away = []
    for team in away_list:
        club, number = extract_club_standings(team, pattern, home=False)
        away.append((club.replace('\t', ''), number))

    # create dataframe with match data
    columns = ["HOME_TEAM", "PLACE_HOME_TEAM", "AWAY_TEAM", "PLACE_AWAY_TEAM","WIN_PERC_HOME", 
               "REMIS_PERC", "WIN_PERC_AWAY"]
    matches = [t1 + t2 + t3 for t1, t2, t3 in zip(home, away, tipps)]
    df = pd.DataFrame(matches, columns=columns)
    
    # retrieve match results
    match_results = soup.find_all('span', class_='matchresult finished')
    result_list = [result.text.strip() for result in match_results]
    
    df['HOME_GOALS'] = [int(element.split(':')[0]) for element in result_list]
    df['AWAY_GOALS'] = [int(element.split(':')[1]) for element in result_list]
    df['RESULT'] = df.apply(lambda row: 'HOME_WIN' if row['HOME_GOALS'] > row['AWAY_GOALS'] else 'AWAY_WIN' if row['HOME_GOALS'] < row['AWAY_GOALS'] else 'DRAW', axis=1)
    
    # retrieve referee for each match
    referees = soup.find_all('a', href = re.compile(r'.*profil/schiedsrichter/.*'))
    df['REFEREE'] = [result.text.strip() for result in referees]
    
    df['DATE'] = dates
    df['DATE'] = pd.to_datetime(df['DATE'], format='%d.%m.%Y')
    df['WEEKDAY'] = df.DATE.dt.weekday
    df['MONTH'] = df.DATE.dt.month
    df['SEASON'] = season
    df['MATCHDAY'] = matchday
    
    # reorder columns
    reorder = ['DATE', 'WEEKDAY', 'MONTH', 'SEASON', 'MATCHDAY']
    df = df[reorder + [col for col in df.columns if col not in reorder]]  
    return df

season_ind_ls

matchday_ind_ls

df_matchday_data = pd.DataFrame()

for season in range(2021, 2022):
    for matchday in matchday_ind_ls:
        df_matchday_data = pd.concat([df_matchday_data, get_matchday_data(season, matchday)])

df_matchday_data.head()

# save dataframe to pickle
#df_matchday_data.to_pickle("./df_matchday_data.pkl")

# read data from pickle
df_matchday_data = pd.read_pickle("./data/df_matchday_data.pkl")

df_matchday_data

# Retrieve transfer data

# Loop through all season for retrieving transfer data
df_transfer_data = pd.DataFrame()
for season in season_ind_ls:
    response = TRANSFERS_URL({season})
    soup = BeautifulSoup(response.content, 'lxml')

    transfer_clubs = soup.find_all('h2', class_='content-box-headline content-box-headline--inverted content-box-headline--logo')
    club_list = [club.text.strip() for club in transfer_clubs]
    club_list


    transfer_infos = soup.find_all('div', class_='transfer-zusatzinfo-box')

    transfer_all = [avg_age.text.strip() for avg_age in transfer_infos]

    # in these seasons there were no "Abgänge" in some clubs
    if season == 1964:
        transfer_all.insert(len(transfer_all), 'Durchschnittsalter der Abgänge: -\t\t\t\t\t\t\t\n\n\t\tGesamtmarktwert der Abgänge: -\t\t\t\t\t\t\t\n\n\t\t\t\t\t\tEinnahmen: 0')
    if  season == 1976:
        transfer_all.insert(17, 'Durchschnittsalter der Abgänge: -\t\t\t\t\t\t\t\n\n\t\tGesamtmarktwert der Abgänge: -\t\t\t\t\t\t\t\n\n\t\t\t\t\t\tEinnahmen: 0')


    transfer_zugaenge = [transfer_all[i] for i in range(len(transfer_all)) if i % 2 == 0]
    transfer_abgaenge = [transfer_all[i] for i in range(len(transfer_all)) if i % 2 == 1]

    # Get average age of joining players
    transfer_avg_age_zugaenge = [string.split('\t\t\t\t\t\t\n\n\t\t\t\t\t')[0] for string in transfer_zugaenge]
    transfer_avg_age_zugaenge_float = [float(string.replace('Durchschnittsalter der Zugänge: ','').replace(',','.')) if ': -' not in string else '' for string in transfer_avg_age_zugaenge]
    transfer_avg_age_zugaenge_float

    # Get average age of leaving players
    transfer_avg_age_abgaenge = [string.split('\t\t\t\t\t\t\t\n\n\t\t')[0] for string in transfer_abgaenge]
    transfer_avg_age_abgaenge_float = [float(string.replace('Durchschnittsalter der Abgänge: ','').replace(',','.')) if ': -' not in string else '' for string in transfer_avg_age_abgaenge]
    transfer_avg_age_abgaenge_float

    # Get total value of joining players
    transfer_value_zugaenge = [string.split('\t\t\t\t\t\t\n\n\t\t\t\t\t')[1].split('\t\t\t\t\t\t\n\n\t\t')[0] for string in transfer_zugaenge]
    transfer_value_zugaenge_float = [float(string.replace('Gesamtmarktwert der Zugänge: ', '').replace(' Mio. €', '').replace(',','.')) if 'Mio.' in string else float(string.replace('Gesamtmarktwert der Zugänge: ', '').replace(' Tsd. €', '').replace(',','.'))/1000 if 'Tsd. €' in string else '' if ': -' not in string else '' for string in transfer_value_zugaenge]
    transfer_value_zugaenge_float

    # Get total value of leaving players
    transfer_value_abgaenge = [string.split('\t\t\t\t\t\t\t\n\n\t\t')[1].split('\t\t\t\t\t\t\t\n\n\t\t\t\t\t\t')[0] for string in transfer_abgaenge]
    transfer_value_abgaenge_float = [float(string.replace('Gesamtmarktwert der Abgänge: ', '').replace(' Mio. €', '').replace(',','.')) if 'Mio.' in string else float(string.replace('Gesamtmarktwert der Abgänge: ', '').replace(' Tsd. €', '').replace(',','.'))/1000 if 'Tsd. €' in string else '' if ': -' not in string else '' for string in transfer_value_abgaenge]
    transfer_value_abgaenge_float

    # Get expenses for joining players
    transfer_ausgaben_zugaenge = [string.split('\t\t\t\t\t\t\n\n\t\t')[2] for string in transfer_zugaenge]
    transfer_ausgaben_zugaenge_float = [float(string.replace('Ausgaben: ', '').replace(' Mio. €', '').replace(',','.')) if 'Mio.' in string else float(string.replace('Ausgaben: ', '').replace(' Tsd. €', '').replace(',','.'))/1000 if 'Tsd. €' in string else '' if ': -' not in string else '' for string in transfer_ausgaben_zugaenge]
    transfer_ausgaben_zugaenge_float

    # Get revenue of leaving players
    transfer_einnahmen_zugaenge = [string.split('\t\t\t\t\t\t\t\n\n\t\t\t\t\t\t')[1] for string in transfer_abgaenge]
    transfer_einnahmen_abgaenge_float = [float(string.replace('Einnahmen: ', '').replace(' Mio. €', '').replace(',','.')) if 'Mio.' in string else float(string.replace('Einnahmen: ', '').replace(' Tsd. €', '').replace(',','.'))/1000 if 'Tsd. €' in string else '' if ': -' not in string else '' for string in transfer_einnahmen_zugaenge]
    transfer_einnahmen_abgaenge_float



    # Write all Information to a Dataframe
    col_name_ls = ['CLUB_NAME','AVG_AGE_JOINING', 'AVG_AGE_LEAVING', 'TOTAL_VALUE_JOINING_MIO', 'TOTAL_VALUE_LEAVING_MIO', 'EXPENSES_JOINING_MIO', 'REVENUE_LEAVING_MIO']
    value_ls = [club_list, transfer_avg_age_zugaenge_float, transfer_avg_age_abgaenge_float, transfer_value_zugaenge_float, transfer_value_abgaenge_float, transfer_ausgaben_zugaenge_float, transfer_einnahmen_abgaenge_float]
    df = dict_to_df(col_name_ls=col_name_ls, value_ls=value_ls) 

    # add season information as column to dataframe
    df['season'] = f"{season}"

    # Concat all dataframes of each season into one dataframe 
    df_transfer_data = pd.concat([df_transfer_data, df])


# save dataframe to pickle
#df_transfer_data.to_pickle("./df_transfer_data.pkl")

# read data from pickle
df_transfer_data = pd.read_pickle("./data/df_transfer_data.pkl")

df_transfer_data
