==> 02_Merging.ipynb <==
import pandas as pd
START_SEASON = 2004
#START_SEASON = 2010

OUTPUT_FILE_NAME = "./data/df_merge_club_transfer_data.pkl"
# OUTPUT_FILE_NAME = "df_merge_club_transfer_data_weather.pkl"
# OUTPUT_FILE_NAME = "df_merge_club_transfer_data_2010.pkl"

df_clubs_overview = pd.read_pickle("./data/df_clubs_overview.pkl")
df_matchday_data = pd.read_pickle("./data/df_matchday_data.pkl")
# df_matchday_data = pd.read_pickle("./df_matchday_data_incl_weather.pkl")
df_transfer_data = pd.read_pickle("./data/df_transfer_data.pkl")

# Merging all dataframes
First renaming clubs, then merge club data with the matchday data

df_clubs_overview['season'] = df_clubs_overview['season'].astype(int)
club_df = df_clubs_overview[df_clubs_overview['season'] >= START_SEASON]
club_df.CLUB_NAME.unique()

nan_count = club_df.isna().sum()
print(nan_count)
print(len(club_df))

df_matchday_data['SEASON'] = df_matchday_data['SEASON'].astype(int)
match_df = df_matchday_data[df_matchday_data['SEASON'] >= START_SEASON]
match_df.HOME_TEAM.unique()

nan_count = match_df.isna().sum()
print(nan_count)
print(len(match_df))

df_transfer_data['season'] = df_transfer_data['season'].astype(int)
transfer_df = df_transfer_data[df_transfer_data['season'] >= START_SEASON]
transfer_df.CLUB_NAME.unique()

nan_count = transfer_df.isna().sum()
print(nan_count)
print(len(transfer_df))

club_mappings = {
   'Werder Bremen': 'SV Werder Bremen',
    "1.FC K'lautern": '1.FC Kaiserslautern',
    'SC Freiburg': 'SC Freiburg',
    'Bor. Dortmund': 'Borussia Dortmund',
    'B. Leverkusen': 'Bayer 04 Leverkusen',
    'Hertha BSC': 'Hertha BSC',
    'Hamburger SV': 'Hamburger SV',
    'Arm. Bielefeld': 'Arminia Bielefeld',
    'VfB Stuttgart': 'VfB Stuttgart',
    '1.FSV Mainz 05': '1.FSV Mainz 05',
    'Bayern München': 'FC Bayern München',
    "Bor. M'gladbach": 'Borussia Mönchengladbach',
    'Hansa Rostock': 'FC Hansa Rostock',
    'FC Schalke 04': 'FC Schalke 04',
    '1.FC Nürnberg': '1.FC Nürnberg',
    'VfL Bochum': 'VfL Bochum',
    'VfL Wolfsburg': 'VfL Wolfsburg',
    'Hannover 96': 'Hannover 96',
    'MSV Duisburg': 'MSV Duisburg',
    '1.FC Köln': '1.FC Köln',
    'E. Frankfurt': 'Eintracht Frankfurt',
    'Energie Cottbus': 'FC Energie Cottbus',
    'Alem. Aachen': 'Alemannia Aachen',
    'Karlsruher SC': 'Karlsruher SC',
    'TSG Hoffenheim': 'TSG 1899 Hoffenheim',
    'FC St. Pauli': 'FC St. Pauli',
    'FC Augsburg': 'FC Augsburg',
    'Greuther Fürth': 'SpVgg Greuther Fürth',
    'F. Düsseldorf': 'Fortuna Düsseldorf',
    'E. Braunschweig': 'Eintracht Braunschweig',
    'SC Paderborn': 'SC Paderborn 07',
    'Darmstadt 98': 'SV Darmstadt 98',
    'FC Ingolstadt': 'FC Ingolstadt 04',
    'RB Leipzig': 'RasenBallsport Leipzig',
    'Union Berlin': '1.FC Union Berlin'
}

# renaming club names in the matchday data
match_df = match_df.copy()
match_df['HOME_TEAM'] = match_df['HOME_TEAM'].replace(club_mappings)
match_df['AWAY_TEAM'] = match_df['AWAY_TEAM'].replace(club_mappings)

# Anwendung des Mappings in transfer_df
transfer_df = transfer_df.copy()
transfer_df['CLUB_NAME'] = transfer_df['CLUB_NAME'].replace(club_mappings)

# merge club data with transfer data for each club and season
tmp_df = pd.merge(club_df, transfer_df, on=['CLUB_NAME', 'season'])

nan_count = tmp_df.isna().sum()
print(nan_count)
print(len(tmp_df))

set_club_transfer_teams = set(tmp_df['CLUB_NAME'])
set_match_df_teams = set(match_df['HOME_TEAM']).union(set(match_df['AWAY_TEAM']))
are_equal = set_club_transfer_teams == set_match_df_teams
print(f"The columns are equal: {are_equal}")

print(set_club_transfer_teams)
print(len(set_club_transfer_teams))

print(set_match_df_teams)
print(len(set_match_df_teams))

# Teams in set1 but not in set2
difference1 = set_club_transfer_teams - set_match_df_teams

# Teams in set2 but not in set1
difference2 = set_match_df_teams - set_club_transfer_teams

# Combine both differences
all_differences = difference1.union(difference2)

print(f"Teams in set1 but not in set2: {difference1}")
print(f"Teams in set2 but not in set1: {difference2}")
print(f"All differing teams: {all_differences}")
print(len(all_differences))

# merge matchday data for home team
tmp2_df = match_df.merge(tmp_df, left_on=['HOME_TEAM', 'SEASON'], right_on=['CLUB_NAME', 'season'], how='left')
tmp2_df = tmp2_df.rename(columns={
    'PLAYERS_COUNT': 'HOME_PLAYERS_COUNT',
    'PLAYERS_AVG_AGE': 'HOME_PLAYERS_AVG_AGE',
    'LEGIONARIES_COUNT': 'HOME_LEGIONARIES_COUNT',
    'AVG_MARKET_VALUE': 'HOME_AVG_MARKET_VALUE',
    'TOTAL_MARKET_VALUE': 'HOME_TOTAL_MARKET_VALUE',
    'AVG_AGE_JOINING': 'HOME_AVG_AGE_JOINING', 
    'AVG_AGE_LEAVING': 'HOME_AVG_AGE_LEAVING', 
    'TOTAL_VALUE_JOINING_MIO': 'HOME_TOTAL_VALUE_JOINING_MIO', 
    'TOTAL_VALUE_LEAVING_MIO': 'HOME_TOTAL_VALUE_LEAVING_MIO', 
    'EXPENSES_JOINING_MIO': 'HOME_EXPENSES_JOINING_MIO',
    'REVENUE_LEAVING_MIO': 'HOME_REVENUE_LEAVING_MIO'
})
tmp2_df

nan_count = tmp2_df.isna().sum()
print(nan_count)

# merge matchday data for away team
df = tmp2_df.merge(tmp_df, left_on=['AWAY_TEAM', 'SEASON'], right_on=['CLUB_NAME', 'season'], how='left')
df = df.rename(columns= {
    'PLAYERS_COUNT': 'AWAY_PLAYERS_COUNT',
    'PLAYERS_AVG_AGE': 'AWAY_PLAYERS_AVG_AGE',
    'LEGIONARIES_COUNT': 'AWAY_LEGIONARIES_COUNT',
    'AVG_MARKET_VALUE': 'AWAY_AVG_MARKET_VALUE',
    'TOTAL_MARKET_VALUE': 'AWAY_TOTAL_MARKET_VALUE',
    'AVG_AGE_JOINING': 'AWAY_AVG_AGE_JOINING', 
    'AVG_AGE_LEAVING': 'AWAY_AVG_AGE_LEAVING', 
    'TOTAL_VALUE_JOINING_MIO': 'AWAY_TOTAL_VALUE_JOINING_MIO', 
    'TOTAL_VALUE_LEAVING_MIO': 'AWAY_TOTAL_VALUE_LEAVING_MIO', 
    'EXPENSES_JOINING_MIO': 'AWAY_EXPENSES_JOINING_MIO',
    'REVENUE_LEAVING_MIO': 'AWAY_REVENUE_LEAVING_MIO'
})
df = df.drop(['CLUB_NAME_x', 'CLUB_NAME_y', 'season_x', 'season_y'], axis=1)

nan_count = df.isna().sum()
print(nan_count)

# save  merge club data with transfer data for each club and season
df.to_pickle(OUTPUT_FILE_NAME)
