==> 03_Prepare.ipynb <==
# Prepare data and convert data types

import pandas as pd
from sklearn.preprocessing import LabelEncoder
OUTPUT_FILE_NAME = "./data/df_merge_club_transfer_data_prep1.pkl"
# OUTPUT_FILE_NAME = "df_merge_club_transfer_data_weather_prep1.pkl"
# OUTPUT_FILE_NAME = "df_merge_club_transfer_data_with_leave_join_mio_prep1.pkl"
DROP_NAN = True

df = pd.read_pickle("./data/df_merge_club_transfer_data.pkl")
#df = pd.read_pickle("./df_merge_club_transfer_data_weather.pkl")

print(df.dtypes)

nan_count = df.isna().sum()
print(nan_count)
print(len(df))

## 1st step of data preparation

df1 = df.copy()
object_columns = df1.select_dtypes(include=['object']).columns
print(object_columns)

print(df1[object_columns].iloc[0])

columns_to_convert = ['PLACE_HOME_TEAM','PLACE_AWAY_TEAM','HOME_PLAYERS_COUNT','HOME_PLAYERS_AVG_AGE','HOME_LEGIONARIES_COUNT',
                     'HOME_AVG_AGE_LEAVING', "HOME_TOTAL_VALUE_JOINING_MIO", "HOME_TOTAL_VALUE_LEAVING_MIO", 'HOME_EXPENSES_JOINING_MIO','HOME_REVENUE_LEAVING_MIO', 
                     'AWAY_PLAYERS_COUNT','AWAY_PLAYERS_AVG_AGE','AWAY_LEGIONARIES_COUNT','AWAY_AVG_AGE_LEAVING', "AWAY_TOTAL_VALUE_JOINING_MIO", "AWAY_TOTAL_VALUE_LEAVING_MIO",
                     'AWAY_EXPENSES_JOINING_MIO', 'AWAY_REVENUE_LEAVING_MIO']

# with weather
# columns_to_convert = ['PLACE_HOME_TEAM','PLACE_AWAY_TEAM','HOME_PLAYERS_COUNT','HOME_PLAYERS_AVG_AGE','HOME_LEGIONARIES_COUNT',
#                      'HOME_AVG_AGE_LEAVING', "HOME_TOTAL_VALUE_JOINING_MIO", "HOME_TOTAL_VALUE_LEAVING_MIO", 'HOME_EXPENSES_JOINING_MIO','HOME_REVENUE_LEAVING_MIO', 
#                      'AWAY_PLAYERS_COUNT','AWAY_PLAYERS_AVG_AGE','AWAY_LEGIONARIES_COUNT','AWAY_AVG_AGE_LEAVING', "AWAY_TOTAL_VALUE_JOINING_MIO", "AWAY_TOTAL_VALUE_LEAVING_MIO",
#                      'AWAY_EXPENSES_JOINING_MIO', 'AWAY_REVENUE_LEAVING_MIO', 'SNOW_HEIGHT_cm', 'RAIN_l/m2', 'SUNSHINE_DURATION_h', 'MIN_TEMP_C', 'MAX_TEMP_C']

for column in columns_to_convert:
    if df1[column].dtype == 'object':  # Nur Spalten vom Typ 'object' verarbeiten
        df1[column] = df1[column].astype(str).str.replace(',', '.')
        df1[column] = pd.to_numeric(df1[column], errors='coerce')

nan_count = df1.isna().sum()
print(nan_count)


first_nan_index = df1[df1['AWAY_EXPENSES_JOINING_MIO'].isna()].index.max()
print(first_nan_index)
value = df.loc[first_nan_index, 'AWAY_EXPENSES_JOINING_MIO']
print(df.loc[first_nan_index, 'AWAY_EXPENSES_JOINING_MIO'])

# Ein Set für die Teamnamen erstellen
unique_teams = set()

# Durch alle Zeilen im DataFrame iterieren und die Teamnamen in das Set aufnehmen, falls "AWAY_EXPENSES_JOINING_MIO" einen bestimmten Wert hat
for index, row in df.iterrows():
    if pd.notna(row['AWAY_EXPENSES_JOINING_MIO']):  # pd.notna prüft, ob der Wert nicht NaN ist
        unique_teams.add(row['AWAY_TEAM'])

# Das Set enthält jetzt die einzigartigen Teamnamen, wo "AWAY_EXPENSES_JOINING_MIO" nicht NaN ist
print(unique_teams)

def convert_market_value(value):
    try:
        # Entferne Kommas und ersetze sie durch Punkte
        value = value.replace(",", ".")
        
        # Millionen
        if "Mio. €" in value:
            value = value.replace("Mio. €", "")
            return float(value) * 1e6  # Multipliziere mit 1 Million

        # Tausend
        elif "Tsd. €" in value:
            value = value.replace("Tsd. €", "")
            return float(value) * 1e3  # Multipliziere mit 1 Tausend

        else:
            return None  # Falls die Konvertierung fehlschlägt, gebe None zurück

    except Exception as e:
        print(f"Konvertierungsfehler: {e}, Wert: {value}")
        return None  # Falls die Konvertierung fehlschlägt, gebe None zurück

df1 = df1.copy()
df1['HOME_AVG_MARKET_VALUE'] = df1['HOME_AVG_MARKET_VALUE'].apply(convert_market_value)
df1['HOME_TOTAL_MARKET_VALUE'] = df1['HOME_TOTAL_MARKET_VALUE'].apply(convert_market_value)
df1['AWAY_AVG_MARKET_VALUE'] = df1['AWAY_AVG_MARKET_VALUE'].apply(convert_market_value)
df1['AWAY_TOTAL_MARKET_VALUE'] = df1['AWAY_TOTAL_MARKET_VALUE'].apply(convert_market_value)
print(df1.dtypes)

print(df1[object_columns].iloc[0])

nan_count = df1.isna().sum()
print(nan_count)

if DROP_NAN == True:
    all_columns = df1.columns.tolist()
    columns_to_exclude = ['WIN_PERC_HOME', 'REMIS_PERC', 'WIN_PERC_AWAY']
    columns_to_include = [col for col in all_columns if col not in columns_to_exclude]
    df1.dropna(subset=columns_to_include, inplace=True)
    df1.reset_index(drop=True)
    
nan_count = df1.isna().sum()
print(nan_count)

df1

df1.to_pickle(OUTPUT_FILE_NAME)
