==> 07_Hypothesis_testing.ipynb <==
# Nullhyptothesis testing. Prediction Model comparison to Random Guess

import pandas as pd
import random

## Read prepared data

pd.set_option('display.max_columns', None)

df = pd.read_pickle("./data/df_merge_club_transfer_data_prep1.pkl")
df_prep = df.copy()
df_prep

## Count distribution of Match Outcomes

result_counts = df['RESULT'].value_counts()
result_percentages = round((result_counts / len(df)), 2)
result_percentages

df_test_data_with_predictions = pd.read_pickle("./data/df_test_data_with_predictions.pkl")
df_test_data_with_predictions

## Apply Random Guess based on the distribution outcome as weights

random.seed(142151)

outcomes = [0, 1, 2]
weights = [result_percentages['AWAY_WIN'], result_percentages['DRAW'], result_percentages['HOME_WIN']]
num_picks = len(df_test_data_with_predictions)

selected_outcomes = random.choices(outcomes, weights=weights, k=num_picks)
selected_outcomes

df_test_data_with_predictions['random_guess'] = selected_outcomes

df_test_data_with_predictions

## Apply McNemar Test to each Model compared to random Guess

model_list = ['random_guess', 'pred_random_forest', 'pred_kneighbor', 'pred_multinomial_nb', 'pred_gaussian_nb', 'pred_quadr_discr_analysis']

for pred in model_list:
    # Flag if model and random guess was classified correctly or not
    df_test_data_with_predictions[f'class_result_{pred}'] = ['Correct' if pred == true else 'Wrong' for pred, true in zip(df_test_data_with_predictions[pred], df_test_data_with_predictions['true_result'])]
    # Create a Contingency table for each model compared to random guess
    contingency_table = pd.crosstab(df_test_data_with_predictions[f'class_result_{pred}'], df_test_data_with_predictions['class_result_random_guess'])
    # Calculate testing statistic
    chi2 = (((contingency_table.iloc[0,1]) - (contingency_table.iloc[1,0]))**2) / ((contingency_table.iloc[0,1])+ (contingency_table.iloc[1,0]))
    print('')
    # Comparison with the Chi^2 distribution
    if chi2 > 3.84:  # 3.84 entspricht dem kritischen Wert der \chi^2 Verteilung mit 1 Freiheitsgrad und 95%-Quantil zum Signifikanzniveau alpha = 0.05 
        print(f'Nullhypothese wird abgelehnt: Es gibt einen signifikanten Unterschied zwischen dem Modell {pred.replace("pred_", "").upper()} und dem Zufallsmodell')
    else:
        print(f'Nullhypothese wird nicht abgelehnt: Es gibt KEINEN signifikanten Unterschied zwischen dem Modell {pred.replace("pred_", "").upper()} und dem Zufallsmodell')
